Refactor merge context to match current common function layout
