+++ examples/default/context/README.md
@@ -10,8 +10,7 @@ $ go run . --insecure --debug
 
 ```shell
 # Then, in another terminal, call it with these example manifests
-$ crossplane render --verbose xr.yaml composition.yaml functions.yaml -r --extra-resources extra_resources.yaml
----
+$ crossplane render --verbose xr.yaml composition.yaml functions.yaml -rc
 ---
 apiVersion: example.crossplane.io/v1beta1
 kind: XR
@@ -20,43 +19,14 @@ metadata:
 status:
   conditions:
   - lastTransitionTime: "2024-01-01T00:00:00Z"
-    message: 'Unready resources: another-awesome-dev-bucket, my-awesome-dev-bucket'
-    reason: Creating
-    status: "False"
+    reason: Available
+    status: "True"
     type: Ready
 ---
-apiVersion: example/v1alpha1
-kind: Foo
-metadata:
-  annotations:
-    crossplane.io/composition-resource-name: another-awesome-dev-bucket
-  generateName: example-
-  labels:
-    crossplane.io/composite: example
-  name: another-awesome-dev-bucket
-  ownerReferences:
-  - apiVersion: example.crossplane.io/v1beta1
-    blockOwnerDeletion: true
-    controller: true
-    kind: XR
-    name: example
-    uid: ""
----
-apiVersion: example/v1alpha1
-kind: Foo
-metadata:
-  annotations:
-    crossplane.io/composition-resource-name: my-awesome-dev-bucket
-  generateName: example-
-  labels:
-    crossplane.io/composite: example
-  name: my-awesome-dev-bucket
-  ownerReferences:
-  - apiVersion: example.crossplane.io/v1beta1
-    blockOwnerDeletion: true
-    controller: true
-    kind: XR
-    name: example
-    uid: ""
-
+apiVersion: render.crossplane.io/v1beta1
+fields:
+  contextField: contextValue
+  moreComplexField:
+    test: field
+kind: Context
 ```
